+++ CHANGELOG.md
@@ -57,3 +57,4 @@
 - [x] plugin_config_id in route 2024-04-28
 - [x] attach the consumer 2024-04-29
 - [x] plugin consumer-restriction [consumer-restriction](https://apisix.apache.org/zh/docs/apisix/plugins/consumer-restriction/) 2024-04-29
+- [x] plugin http-logger [http-logger](https://apisix.apache.org/zh/docs/apisix/plugins/http-logger/) 2024-04-30